+++ packages/picimo/src/hooks/useTextureAtlas.ts
@@ -1,25 +1,25 @@
 import {TextureAtlas} from '@spearwolf/vertex-objects';
 import {useContext, useEffect, useState} from 'react';
-import {AssetName, TextureStoreContext} from '../context/TextureStore';
+import {AssetName, AssetStoreContext} from '../context/AssetStore';
 
 export function useTextureAtlas(name: AssetName): TextureAtlas | undefined {
-  const textureStore = useContext(TextureStoreContext);
+  const assetStore = useContext(AssetStoreContext);
   const [curTextureAtlas, setTextureAtlas] = useState<TextureAtlas>();
 
   useEffect(() => {
-    const atlas = textureStore.getTextureAtlas(name);
+    const atlas = assetStore.getTextureAtlas(name);
     if (atlas) {
       setTextureAtlas(atlas);
     }
-    return textureStore.on('asset:insert', (assetName: AssetName) => {
+    return assetStore.on('asset:insert', (assetName: AssetName) => {
       if (name === assetName) {
-        const nextAtlas = textureStore.getTextureAtlas(name);
+        const nextAtlas = assetStore.getTextureAtlas(name);
         if (nextAtlas !== curTextureAtlas) {
           setTextureAtlas(nextAtlas);
         }
       }
     });
-  }, [textureStore, name]);
+  }, [assetStore, name]);
 
   return curTextureAtlas;
 }